Move recurring-event-picker to its own directory
Mainly to add a custom css file
